Split off EPD Test
