+++ epdtest/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+version = "1.0.0"
+name = "epdtest"
+authors.workspace = true
+edition.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+librustic = { path = "../librustic" }
