+++ epdtest/src/main.rs
@@ -1,6 +1,8 @@
-use crate::{
+mod epds;
+
+use epds::LARGE_TEST_EPDS;
+use librustic::{
     board::Board,
-    extra::epds::LARGE_TEST_EPDS,
     misc::perft,
     movegen::MoveGenerator,
     search::defs::{PerftData, TT},
@@ -29,7 +31,10 @@ const TEST_RESULTS: [&str; 5] = [
 
 // This private function is the one actually running tests.
 // This can be the entire suite, or a single test.
-pub fn run(tt_size: usize) {
+pub fn main() {
+    // TODO: Add command-line to set transposition table size
+
+    let tt_size = 0;
     let number_of_tests = LARGE_TEST_EPDS.len();
     let move_generator = MoveGenerator::new();
     let mut board: Board = Board::new();
