--- librustic/src/extra.rs
@@ -1,2 +1,0 @@
-pub mod epds;
-pub mod testsuite;
